+++ catalystwan/endpoints/configuration/policy/list/url_block_list.py
@@ -5,17 +5,9 @@ from uuid import UUID
 
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
 from catalystwan.endpoints.configuration.policy.abstractions import PolicyListEndpoints
-from catalystwan.models.policy.lists import URLBlockList
-from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListInfo, PolicyListPreview
+from catalystwan.models.policy.list.url import URLBlockList, URLBlockListEditPayload, URLBlockListInfo
+from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListPreview
 from catalystwan.typed_list import DataSequence
-
-
-class URLBlockListEditPayload(URLBlockList, PolicyListId):
-    pass
-
-
-class URLBlockListInfo(URLBlockList, PolicyListInfo):
-    pass
 
 
 class ConfigurationPolicyURLBlockList(APIEndpoints, PolicyListEndpoints):
